Make a HeadlessBrowser abstraction and move it to a separate file
