+++ lib/main.js
@@ -5,92 +5,7 @@ const groupBy = require('lodash.groupby');
 const urlTools = require('urltools');
 const util = require('util');
 const subsetFonts = require('./subsetFonts');
-
-async function transferResults(jsHandle) {
-  const results = await jsHandle.jsonValue();
-  for (const [i, result] of results.entries()) {
-    const resultHandle = await jsHandle.getProperty(String(i));
-    const elementHandle = await resultHandle.getProperty('node');
-    result.node = elementHandle;
-  }
-  return results;
-}
-
-async function downloadAndLaunchHeadlessBrowser() {
-  const puppeteer = require('puppeteer-core');
-  const browserFetcher = puppeteer.createBrowserFetcher();
-  const preferredRevision = puppeteer._launcher._preferredRevision;
-  const localRevisions = await browserFetcher.localRevisions();
-  let revisionInfo;
-  if (localRevisions.includes(preferredRevision)) {
-    revisionInfo = await browserFetcher.revisionInfo(preferredRevision);
-  } else {
-    console.log(`Downloading Chromium ${preferredRevision}`);
-    revisionInfo = await browserFetcher.download(preferredRevision);
-  }
-  return await puppeteer.launch({
-    executablePath: revisionInfo.executablePath
-  });
-}
-
-async function tracePageInBrowser(browser, assetGraph, htmlAsset) {
-  const page = await browser.newPage();
-
-  // Make up a base url to map to the assetgraph root.
-  // Use the canonical root if available, so that it'll be
-  // easier to handle absolute and protocol-relative urls pointing
-  // at it, as well as fall through to the actual domain if some
-  // assets aren't found in the graph.
-  const baseUrl = assetGraph.canonicalRoot
-    ? assetGraph.canonicalRoot.replace(/\/?$/, '/')
-    : 'https://example.com/';
-
-  // Intercept all requests made by the headless browser, and
-  // fake a response from the assetgraph instance if the corresponding
-  // asset is found there:
-  await page.setRequestInterception(true);
-  page.on('request', request => {
-    const url = request.url();
-    if (url.startsWith(baseUrl)) {
-      let agUrl = url.replace(baseUrl, assetGraph.root);
-      if (/\/$/.test(agUrl)) {
-        agUrl += 'index.html';
-      }
-      const asset = assetGraph.findAssets({
-        isLoaded: true,
-        url: agUrl
-      })[0];
-      if (asset) {
-        request.respond({
-          status: 200,
-          contentType: asset.contentType,
-          body: asset.rawSrc
-        });
-        return;
-      }
-    }
-    // Not found, let the original request through:
-    request.continue();
-  });
-
-  await page.goto(
-    urlTools.resolveUrl(
-      baseUrl,
-      urlTools.buildRelativeUrl(assetGraph.root, htmlAsset.url)
-    )
-  );
-
-  await page.addScriptTag({
-    path: require.resolve('font-tracer/dist/fontTracer.browser.js')
-  });
-
-  const jsHandle = await page.evaluateHandle(
-    /* global fontTracer */
-    /* istanbul ignore next */
-    () => fontTracer(document)
-  );
-  return transferResults(jsHandle);
-}
+const HeadlessBrowser = require('./HeadlessBrowser');
 
 module.exports = async (argv, console) => {
   const {
@@ -309,7 +224,7 @@ module.exports = async (argv, console) => {
   let tracesByAsset;
   if (dynamic) {
     tracesByAsset = new Map();
-    const browser = await downloadAndLaunchHeadlessBrowser();
+    const headlessBrowser = new HeadlessBrowser();
     try {
       for (const htmlAsset of assetGraph.findAssets({
         type: 'Html',
@@ -317,11 +232,11 @@ module.exports = async (argv, console) => {
       })) {
         tracesByAsset.set(
           htmlAsset,
-          await tracePageInBrowser(browser, assetGraph, htmlAsset)
+          await headlessBrowser.tracePage(htmlAsset)
         );
       }
     } finally {
-      await browser.close();
+      await headlessBrowser.close();
     }
   }
